Throw if attacher in transaction on current thread
Android will throw an exception if you try to attach a database while
the attacher has an open transaction on the current thread. This was
suppressed by tryExecSql and we returned null instead.

The fix to acquire the exclusive lock on API 16+ if WAL is enabled
now causes an exception to be thrown in this situation; if WAL is not
enabled, or on API 15 and below, null was still returned. This change
makes the behavior consistent in all cases.