Hack for handling null characters in inlined string literals
Java treats strings differently than C does. In Java, null characters are allowed. In C, they end the string. So if we inline a Java string that contains null characters as a literal, the parsing will just fail. Of course, the best option is to bind all such strings as arguments rather than inlining them, but when there are too many arguments, we fall back to inlining, so we need a fallback. This commit enhances string literal sanitization to replace occurrences of null characters with SQL string concatenations of CAST(x'00' AS TEXT), which converts a literal blob of the null character to a string.